Cookie

+++ src/main/java/com/example/block/repository/ContestRepository.java
@@ -2,9 +2,8 @@ package com.example.block.repository;
 
 import com.example.block.domain.Contest;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface ContestRepository extends JpaRepository<Contest, Long>{
+public interface ContestRepository extends JpaRepository<Contest, Integer>{
     boolean existsByTitle(String title);
 }
